requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langchain-huggingface
chromadb
pypdf
sentence-transformers
pandas
plotly
faiss-cpu
python-dotenv
flashrank